requirements.txt
msgpack>=1.0.8
pynacl>=1.5.0
cryptography>=42.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
rich>=13.7.0
httpx>=0.27.0